requirements.txt
pandas 
matplotlib
seaborn
torch
torchvision
numpy
torchaudio
transformers
soundfile
scikit-learn
tqdm
protobuf
huggingface_hub
python-dotenv